Criação animalService para função de validar animal para abate e ajustes nos formatos de data e numero

+++ src/Controller/AnimalController.php
@@ -5,6 +5,7 @@ namespace App\Controller;
 use App\Entity\Animal;
 use App\Form\AnimalType;
 use App\Repository\AnimalRepository;
+use App\Service\AnimalService;
 use App\Service\MessageGenerator;
 use DateTime;
 use Doctrine\ORM\EntityManagerInterface;
@@ -15,6 +16,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Security\Http\Attribute\IsGranted;
+
 class AnimalController extends AbstractController
 {
 
@@ -23,23 +25,21 @@ class AnimalController extends AbstractController
     public function index(Request $request, AnimalRepository $animalRepository, PaginatorInterface $paginator): Response
     {
         try {
-            $descricaoanimal = $request->query->get('descricao');
-            //restringir a pagina apenas aos ROLE_USER
-            $this->denyAccessunlessGranted('ROLE_USER');
+            $pesquisadescricaoanimal = $request->query->get('pesquisaDescricao');
             //buscar no bd todos os animais cadastrados
-            $query = is_null($descricaoanimal) #condição ternaria#
+            $query = is_null($pesquisadescricaoanimal) #condição ternaria#
                 ? $animalRepository->findAnimal()
-                : $animalRepository->findAnimalByLikeDescricao($descricaoanimal);
+                : $animalRepository->findAnimalByLikeDescricao($pesquisadescricaoanimal);
             $pagination = $paginator->paginate(
                 $query,
                 $request->query->getInt('page', 1),
                 10
             );
             $data['pagination'] = $pagination;
-            $data['descricao'] = $descricaoanimal;
+            $data['pesquisaDescricao'] = $pesquisadescricaoanimal;
             $data['titulo'] = 'Gerenciar Animais';
 
-            return $this->render('animal/index.html.twig', $data);
+            return $this->render('Animal/index.html.twig', $data);
         } catch (Exception $e) {
             $this->addFlash(
                 'error',
@@ -54,9 +54,6 @@ class AnimalController extends AbstractController
     public function adicionar(Request $request, EntityManagerInterface $em, MessageGenerator $messageGenerator): Response
     {
         try {
-            //restringir a pagina apenas aos ROLE_USER
-            $this->denyAccessunlessGranted('ROLE_USER');
-
             $animal = new Animal();
             $animal->setStatus(0);
             $form = $this->createForm(AnimalType::class, $animal);
@@ -68,7 +65,7 @@ class AnimalController extends AbstractController
                 $em->flush();
                 $message = $messageGenerator->getMessageCreateSuccess();
                 $this->addFlash('notice', $message);
-               
+
                 return $this->redirectToRoute('animal_index');
             }
 
@@ -76,7 +73,7 @@ class AnimalController extends AbstractController
             $data['titulo'] = 'Cadastrar novo animal';
             $data['form'] = $form;
 
-            return $this->renderform('animal/form.html.twig', $data);
+            return $this->render('Animal/form.html.twig', $data);
         } catch (Exception $e) {
             $this->addFlash(
                 'error',
@@ -91,24 +88,28 @@ class AnimalController extends AbstractController
     public function editar($id, Request $request, EntityManagerInterface $em, AnimalRepository $animalRepository, MessageGenerator $messageGenerator): Response
     {
         try {
-            //restringir a pagina apenas aos ROLE_USER
-            $this->denyAccessunlessGranted('ROLE_USER');
-
             $animal = $animalRepository->find($id); //retorna animalpelo $id
-            $form = $this->createForm(AnimalType::class, $animal);
-            $form->handleRequest($request);
+            if ($animal->isStatus() == 0) {
+                $form = $this->createForm(AnimalType::class, $animal);
+                $form->handleRequest($request);
 
-            if ($form->isSubmitted() && $form->isValid()) {
-                $em->flush(); //fazer update do animal no bd
-                $message = $messageGenerator->getMessageUpdateSuccess();
-                $this->addFlash('notice', $message);
-                return $this->redirectToRoute('animal_index');
+                if ($form->isSubmitted() && $form->isValid()) {
+                    $em->flush(); //fazer update do animal no bd
+                    $message = $messageGenerator->getMessageUpdateSuccess();
+                    $this->addFlash('notice', $message);
+                    return $this->redirectToRoute('animal_index');
+                }
+
+                $data['titulo'] = 'Editar animal';
+                $data['form'] = $form;
+
+                return $this->render('Animal/form.html.twig', $data);
+            } else {
+                $message = $messageGenerator->getMessageUpdateError();
+                $this->addFlash('error', $message);
+                
             }
-
-            $data['titulo'] = 'Editar animal';
-            $data['form'] = $form;
-
-            return $this->renderform('animal/form.html.twig', $data);
+            return $this->redirectToRoute('animal_index');
         } catch (Exception $e) {
             $this->addFlash(
                 'error',
@@ -120,18 +121,20 @@ class AnimalController extends AbstractController
 
     #[isGranted('ROLE_USER')]
     #[Route('/animal/excluir/{id}', name: 'animal_excluir')]
-    public function excluir($id, EntityManagerInterface $em, AnimalRepository $animalRepository): Response
+    public function excluir($id, EntityManagerInterface $em, AnimalRepository $animalRepository, MessageGenerator $messageGenerator): Response
     {
         try {
-            //restringir a pagina apenas aos ROLE_USER
-            $this->denyAccessunlessGranted('ROLE_USER');
             $animal = $animalRepository->find($id);
-            $em->remove($animal); //excluir animal do bd
-            $em->flush(); //excluir o animal do bd
-            $this->addFlash(
-                'notice',
-                'Animal excluído com sucesso!'
-            );
+            if ($animal->isStatus() == 0) {
+                $em->remove($animal); //excluir animal do bd
+                $em->flush(); //excluir o animal do bd
+                $message = $messageGenerator->getMessageDeleteSuccess();
+                $this->addFlash('notice', $message);
+            } else {
+                $message = $messageGenerator->getMessageDeleteError();
+                $this->addFlash('error', $message);
+            }
+
 
             return $this->redirectToRoute('animal_index');
         } catch (Exception $e) {
@@ -145,23 +148,20 @@ class AnimalController extends AbstractController
 
     #[isGranted('ROLE_USER')]
     #[Route('/animal/abaterAnimal/{id}', name: 'animal_abater')]
-    public function abaterAnimal($id, EntityManagerInterface $em, AnimalRepository $animalRepository, MessageGenerator $messageGenerator): Response
+    public function abaterAnimal($id, EntityManagerInterface $em, AnimalRepository $animalRepository, MessageGenerator $messageGenerator, AnimalService $animalService): Response
     {
         try {
-            //restringir a pagina apenas aos ROLE_USER
-            $this->denyAccessunlessGranted('ROLE_USER');
-
-            $animais = $animalRepository->findAnimalForAbate();
-            foreach ($animais as $animalAbate) {                
-                if ($animalAbate->getId() == $id) {
-                    $animal = $animalRepository->find($id);
-                    $animal->setStatus(1);
-                    $animal->setDtabate(new DateTime());
-                    $em->flush(); //excluir o animal do bd
-                    $message = $messageGenerator->getMessageAbaterSuccess();
-                    $this->addFlash('notice', $message);
-                } 
-            }           
+            $animal = $animalRepository->find($id);
+            if ($animalService->verificaAnimalAbate($animal)) {
+                $animal->setStatus(1);
+                $animal->setDtabate(new DateTime());
+                $em->flush();
+                $message = $messageGenerator->getMessageAbaterSuccess();
+                $this->addFlash('notice', $message);
+            } else {
+                $message = $messageGenerator->getMessageAbaterError();
+                $this->addFlash('error', $message);
+            }
 
             return $this->redirectToRoute('animal_abate');
         } catch (Exception $e) {
@@ -178,8 +178,6 @@ class AnimalController extends AbstractController
     public function relatorioAbate(Request $request, AnimalRepository $animalRepository, PaginatorInterface $paginator): Response
     {
         try {
-            //restringir a pagina apenas aos ROLE_USER
-            $this->denyAccessunlessGranted('ROLE_USER');
             //buscar no bd todos os animais cadastrados
             $query = $animalRepository->findAnimalForAbate();
             $pagination = $paginator->paginate(
@@ -189,7 +187,7 @@ class AnimalController extends AbstractController
             );
             $data['titulo'] = "Animais para Abater";
             $data['pagination'] = $pagination;
-            return $this->render('animal/relatorio.html.twig', $data);
+            return $this->render('Animal/relatorio.html.twig', $data);
         } catch (Exception $e) {
             $this->addFlash(
                 'error',
@@ -204,8 +202,6 @@ class AnimalController extends AbstractController
     public function relatorioAnimaisAbatidos(Request $request, AnimalRepository $animalRepository, PaginatorInterface $paginator): Response
     {
         try {
-            //restringir a pagina apenas aos ROLE_USER
-            $this->denyAccessunlessGranted('ROLE_USER');
             //buscar no bd todos os animais cadastrados
             $query = $animalRepository->findByStatus(1);
             $pagination = $paginator->paginate(
@@ -215,7 +211,7 @@ class AnimalController extends AbstractController
             );
             $data['titulo'] = "Animais Abatidos";
             $data['pagination'] = $pagination;
-            return $this->render('animal/relatorioAbatidos.html.twig', $data);
+            return $this->render('Animal/relatorioAbatidos.html.twig', $data);
         } catch (Exception $e) {
             $this->addFlash(
                 'error',
